// src/db/schema.d.ts
/**
 * This file was generated by kysely-codegen.
 * Please do not edit it manually.
 */

import type { ColumnType } from "kysely";

export type Generated<T> =
  T extends ColumnType<infer S, infer I, infer U>
    ? ColumnType<S, I | undefined, U>
    : ColumnType<T, T | undefined, T>;

export type Int8 = ColumnType<
  string,
  bigint | number | string,
  bigint | number | string
>;

export type Numeric = ColumnType<string, number | string, number | string>;

export type Timestamp = ColumnType<Date, Date | string, Date | string>;

export interface Address {
  balance: Numeric;
  created_at: Generated<Timestamp>;
  id: string;
  updated_at: Generated<Timestamp>;
}

export interface Block {
  created_at: Generated<Timestamp>;
  height: Int8;
  id: string;
  updated_at: Generated<Timestamp>;
}

export interface Input {
  created_at: Generated<Timestamp>;
  index: number;
  output_reference_index: number;
  output_reference_transaction_id: string;
  transaction_id: string;
  updated_at: Generated<Timestamp>;
}

export interface Output {
  address_id: string;
  created_at: Generated<Timestamp>;
  index: number;
  transaction_id: string;
  updated_at: Generated<Timestamp>;
  value: Numeric;
}

export interface Transactions {
  block_id: string;
  created_at: Generated<Timestamp>;
  id: string;
  updated_at: Generated<Timestamp>;
}

export interface DB {
  address: Address;
  block: Block;
  input: Input;
  output: Output;
  transactions: Transactions;
}